Fix mypy strict mode type errors
- Add scipy-stubs to dev dependencies for proper scipy type hints
- Use .item() for numpy scalar to Python type conversion
- Replace global statement with ModelFactoryManager class pattern
- Add explicit type annotations for defaults.with_updates()
- Fix matplotlib Figure type imports and plt.get_cmap() usage
- Add Callable type annotation for MCP resource handlers
- Update calibration methods to use concrete return types
- Add uv add instructions to CLAUDE.md

Remaining type: ignore comments are for MCP library decorators
pending upstream fix (modelcontextprotocol/python-sdk#1980)

+++ src/japan_fiscal_simulator/core/simulation.py
@@ -234,8 +234,8 @@ class FiscalMultiplierCalculator:
             y_cum = np.sum(y_response[:h])
             policy_cum = np.sum(policy_response[:h])
             if abs(policy_cum) > 1e-10:
-                return y_cum / policy_cum / policy_ratio
-            return y_cum / (0.01 * h) / policy_ratio
+                return float(y_cum / policy_cum / policy_ratio)
+            return float(y_cum / (0.01 * h) / policy_ratio)
 
         cumulative_4q = cumulative_mult(4)
         cumulative_8q = cumulative_mult(8)
